check status on delete class

+++ routes/api/board/controller.js
@@ -88,21 +88,26 @@ exports.delete = function (req, res) {
             console.log ('DB delete err: ', err);
             res.status(500).send('Err: DB delete Error');
         } else {
-            chatModel.deleteByClassNumber(req.params.num, function (err) {
-                if (err) {
+            chatModel.deleteByClassNumber(req.params.num, function (status) {
+                if (status === 500) {
                     res.status(500).send();
                 } else {
-		console.log('chat delete');
-                    exec('docker stop ' + req.params.num + '&&docker rm ' + req.params.num
-                        + '&&rm -rf /root/store/' + req.params.num, function (err) {
-                        if (err) {
-                            console.log ('Docker remove err: ', err);
-                            res.status(500).send();
+                    model.getStatus(req.param.num, function (result) {
+                        if (result.length !== 0) {
+                            exec('docker stop ' + req.params.num + '&&docker rm ' + req.params.num
+                                + '&&rm -rf /root/store/' + req.params.num, function (err) {
+                                if (err) {
+                                    console.log ('Docker remove err: ', err);
+                                    res.status(500).send();
+                                } else {
+                                    res.status(200).send();
+                                }
+                            });
                         } else {
-			console.log('docker delete');
                             res.status(200).send();
                         }
-                    });
+                    })
+
                 }
             });
         }
